Move slicer to package
