+++ xfields/__init__.py
@@ -10,6 +10,8 @@ from .fieldmaps import TriLinearInterpolatedFieldMap
 from .fieldmaps import TriCubicInterpolatedFieldMap
 from .fieldmaps import BiGaussianFieldMap, mean_and_std
 
+from .slicers import UniformBinSlicer
+
 from .solvers.fftsolvers import FFTSolver3D
 
 from .beam_elements.spacecharge import SpaceCharge3D, SpaceChargeBiGaussian
